+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix administrators losing access to a block's configuration after setting a profile to "no access" on that block.
 - Fix dependency conflict with GLPI core by no longer vendoring symfony/deprecation-contracts and symfony/polyfill-ctype.
 - Fix default field values not being applied when fields are empty on creation
+- Fix a field's default value not being applied to existing items and not being shown in search results for items with no dedicated row in the container table
 
 ## [1.24.5] - 2026-09-11